refactor(mf2): Move & rename parsers, now use messageFromCST(parseCST(src)) instead of asDataModel(parseMessage(src))
