+++ packages/mf2-messageformat/src/cst-parser/expression.ts
@@ -1,6 +1,6 @@
 import { MessageSyntaxError } from '../errors.js';
 import type * as CST from './cst-types.js';
-import type { ParseContext } from './message.js';
+import type { ParseContext } from './parse-cst.js';
 import { parseNameValue } from './names.js';
 import { whitespaceChars, whitespaces } from './util.js';
 import { parseLiteral, parseQuotedLiteral, parseVariable } from './values.js';
