--- packages/mf2-messageformat/src/cst-parser/index.ts
@@ -1,4 +1,0 @@
-import type * as CST from './cst-types.js';
-export { CST };
-export { asDataModel, cst } from './as-data-model.js';
-export { parseMessage } from './message.js';
