+++ packages/mf2-messageformat/src/data-model/types.ts
@@ -1,5 +1,5 @@
-import { cst } from '../cst-parser/as-data-model.js';
 import type * as CST from '../cst-parser/cst-types.js';
+import { cst } from './from-cst.js';
 
 /** A node in a message data model */
 export type MessageNode =
