+++ packages/mf2-messageformat/src/stringifier/message.test.ts
@@ -1,12 +1,15 @@
 import testCore from '../__fixtures/test-core.json';
 import testFunctions from '../__fixtures/test-functions.json';
-import { testName } from '../__fixtures/util-test-name';
-import { asDataModel, parseMessage } from '../cst-parser/index.js';
-import { Message } from '../data-model/types';
-import { cst } from '../cst-parser/as-data-model';
-import { visit } from '../data-model/visit';
+import { testName } from '../__fixtures/util-test-name.js';
+import {
+  type Message,
+  cst,
+  messageFromCST,
+  parseCST,
+  stringifyMessage,
+  visit
+} from '../index.js';
 import type { TestMessage } from '../messageformat.test.js';
-import { stringifyMessage } from './message.js';
 
 function trimCST(msg: Message) {
   delete msg[cst];
@@ -28,10 +31,10 @@ for (const [title, messages] of [
     for (const testMsg of messages) {
       if (!testMsg.errors?.length) {
         test(testName(testMsg), () => {
-          const msg0 = asDataModel(parseMessage(testMsg.src));
+          const msg0 = messageFromCST(parseCST(testMsg.src));
           trimCST(msg0);
           const src1 = stringifyMessage(msg0);
-          const msg1 = asDataModel(parseMessage(src1));
+          const msg1 = messageFromCST(parseCST(src1));
           trimCST(msg1);
           expect(msg1).toEqual(msg0);
         });
